[fix] HttpStatusException을 GlobalExceptionHandler에서 처리
